requirements.txt
flask
twilio
openai
anthropic
pyyaml 
python-dotenv 
flask_sqlalchemy 
flask_migrate
gunicorn
colorama
flask_jwt_extended
werkzeug
celery
redis
uvicorn
fastapi
jinja2
python-multipart
itsdangerous
aiofiles
starlette
asyncpg
greenlet
python-jose
passlib
pydantic
playwright
pandas 
pgvector 
psycopg
pytest
psycopg2-binary
pinecone
langchain
